+++ backend/src/service/commands/mod.rs
@@ -1,0 +1,6 @@
+pub mod ffmpeg;
+mod ffprobe;
+mod yt_dlp;
+
+pub use self::ffprobe::ffprobe;
+pub use self::yt_dlp::{YtDlp, YtDlpOptions};