+++ format.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
+cd backend
 cargo fix --allow-staged
 cargo +nightly fmt
-cd frontend
+cd ../frontend
 npm run format
 cd ..