+++ frontend/vite.config.ts
@@ -1,13 +1,13 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react-swc'
-import eslint from 'vite-plugin-eslint'
+import {defineConfig} from "vite"
+import react from "@vitejs/plugin-react-swc"
+import eslint from "vite-plugin-eslint"
 
 // https://vitejs.dev/config/
 export default defineConfig({
   server: {
     proxy: {
-      "/api": "http://localhost:5174"
-    }
+      "/api": "http://localhost:5174",
+    },
   },
   plugins: [react(), eslint({})],
 })