+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## Unreleased
+- feat: Allow downloading videos via `yt-dlp`
+- feat: Allow editing the marker start/end time fields
+- feat: Validate markers before saving them to the database
+- fix: Beats were not being generated for music
+- chore: move backend code into separate folder
 
 ## 0.8.0
 - feat: Generate clips based on a song's beats